requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Faker==19.2.0
fastapi==0.95.1
frozenlist==1.4.1
grab-favicon==1.0.9
greenlet==2.0.2
h11==0.14.0
idna==3.4
libsql-client==0.3.0
multidict==6.0.4
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
sqlalchemy-libsql==0.1.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.4
uvicorn==0.21.1
yarl==1.9.4
